requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.0.6
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
openai>=1.0.0
httpx>=0.23.0,<1.0.0
pytest>=6.2.4,<6.3.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
pillow>=9.0.0
numpy>=1.21.0
pytest-asyncio>=0.18.0
